Refactor associations… again
This is a lot better.  Basically what it comes down to is that the :conditions option is only available on has_many associations and NOT on has_many :through. The solution is to create a couple of specialty HM associations and use HM:T on those.  THis allows a clear path way through the association that will automatically set the artistic_role_id attribute.