Completed youtube support.
Youtube support is done via a rake task.  The rake task will search for videos titled similarly to "owning artist name - track name".  The search limits to embeddable videos (aka format: 5 in the youtube api).

This task needs to be run periodically or as videos are removed the pages may break.